ConstraintCommons

Documentation for ConstraintCommons.

ConstraintCommons.USUAL_CONSTRAINT_PARAMETERS

ConstraintCommons.AbstractAutomaton

ConstraintCommons.AbstractMultivaluedDecisionDiagram

ConstraintCommons.Automaton

ConstraintCommons.Automaton

ConstraintCommons.MDD

Base.:*

Base.in

Base.isempty

ConstraintCommons.accept

ConstraintCommons.at_end

ConstraintCommons.extract_parameters

ConstraintCommons.incsert!

ConstraintCommons.oversample

ConstraintCommons.δ_extrema

ConstraintCommons.USUAL_CONSTRAINT_PARAMETERS — Constant

const USUAL_CONSTRAINT_PARAMETERS

List of usual constraints parameters (based on XCSP3-core constraints).

const USUAL_CONSTRAINT_PARAMETERS = [
    :bool,
    :dim,
    :id,
    :language,
    :op,
    :pair_vars,
    :val,
    :vals,
]

ConstraintCommons.AbstractAutomaton — Type

AbstractAutomaton

An abstract interface for automata used in Julia Constraints packages. Requirements:

accept(a<:AbstractAutomaton, word): return true if a accepts word.

ConstraintCommons.AbstractMultivaluedDecisionDiagram — Type

AbstractMultivaluedDecisionDiagram

An abstract interface for Multivalued Decision Diagrams (MDD) used in Julia Constraints packages. Requirements:

accept(a<:AbstractMultivaluedDecisionDiagram, word): return true if a accepts word.

ConstraintCommons.Automaton — Type

Automaton{S, T, F <: Union{S, Vector{S}, Set{S}}} <: AbstractAutomaton

A minimal implementation of a deterministic automaton structure.

ConstraintCommons.Automaton — Method

Automaton(a::MDD)

Construct an automaton based on a given Multivalued Decision Diagrams (MDD).

ConstraintCommons.MDD — Type

MDD{S,T} <: AbstractMultivaluedDecisionDiagram

A minimal implementation of a multivalued decision diagram structure.

Base.:* — Function

Base.:*(s1::Symbol, s2::Symbol, connector::AbstractString="_")

Extends * to Symbols multiplication by connecting the symbols by an _.

Base.in — Method

Base.in(::Any, ::Nothing)

Extends Base.in (or ∈) when the set is nothing. Returns false.

Base.isempty — Method

Base.isempty(::Nothing)

Extends Base.isempty when the set is nothing. Returns true.

ConstraintCommons.accept — Method

accept(a::Union{Automaton, MDD}, w)

Return true if a accepts the word w and false otherwise.

ConstraintCommons.at_end — Method

at_end(a::Automaton, s)

Internal method used by accept with Automaton.

ConstraintCommons.extract_parameters — Method

extract_parameters(m::Union{Method, Function}; parameters)

Extracts the intersection between the kargs of m and parameters (defaults to USUAL_CONSTRAINT_PARAMETERS).

ConstraintCommons.incsert! — Function

incsert!(d::Union{AbstractDict, AbstractDictionary}, ind, val = 1)

Increase or insert a counter in a dictionary-based collection. The counter insertion defaults to val = 1.

ConstraintCommons.oversample — Method

oversample(X, f)

Oversample elements of X until the boolean function f has as many true and false configurations.

ConstraintCommons.δ_extrema — Method

δ_extrema(X...)

Compute both the difference between the maximum and the minimum of over all the collections of X.